requirements.txt
django==4.2
django-environ==0.11.2
psycopg2-binary==2.9.9
pillow==10.4.0
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-modeltranslation==0.19.5
djangorestframework==3.15.2
drf-yasg==1.21.7
setuptools==70.3.0
django-jazzmin==3.0.0
django-redis==5.4.0
django-filter==24.2
celery==5.4.0
djangorestframework-simplejwt==5.3.1